oct-2128: home e2e
