+++ client/cypress/e2e/_02makePendingSnapshot.cy.ts
