+++ client/cypress/e2e/_05layoutFooter.cy.ts
