+++ client/cypress/e2e/_06layoutNavbar.cy.ts
