+++ client/cypress/e2e/_07onboardingTOSNotAccepted.cy.ts
