+++ client/cypress/e2e/_08onboardingTOSAccepted.cy.ts
